requirements.txt
social-auth-app-django
djangorestframework
django-mptt
pytils